Removed CharacterInfo and refactored into Character and CharacterHandler
